add authentication token, reset password, display flash message, flask-moment, flask-mail

+++ app/course/views.py
@@ -1,17 +1,12 @@
-from flask import render_template, redirect, url_for, flash
+from flask import render_template, redirect, url_for, flash, abort
 from . import course
 from flask_login import login_required
 from .forms import AddCourseForm
-from ..models import Course
+from ..models import Course, Journal
 from app import db
 
-# @course.route('/index')
-# def index():
-#     form = AddCourseForm()
-#     courses = Course.query.all()
-#     render_template('course/index.html', courses=courses, form=form)
 
-@course.route('/index', methods=['GET','POST'])
+@course.route('/', methods=['GET','POST'])
 @login_required
 def index():
     form = AddCourseForm()
@@ -20,7 +15,7 @@ def index():
                         semester=form.semester.data)
         db.session.add(course)
         db.session.commit()
-        flash('Course registered')
+        flash('Course registered successfully','success')
         return redirect(url_for('course.index'))
     courses = Course.query.order_by(Course.semester).all()
     return render_template('course/index.html', form=form, courses=courses)
@@ -30,7 +25,10 @@ def index():
 @login_required
 def remove(id):
     course = Course.query.filter_by(id=int(id)).first()
-    db.session.delete(course)
-    db.session.commit()
-    flash('Course has been removed')
+    if len(course.journals) != 0: # check if this course is connected to some journals.
+        flash("you are not allowed to delete this course, because this course is being used by other students.", 'warning')
+    else:
+        db.session.delete(course)
+        db.session.commit()
+        flash('Course has been removed','info')
     return redirect(url_for('course.index'))
